rename of individual_sub_answer_generation
