+++ backend/onyx/agents/agent_search/deep_search_a/initial/sub_answer_consolidation/nodes/ingest_initial_sub_answers.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from onyx.agents.agent_search.deep_search_a.initial.individual_sub_answer_generation.states import (
+from onyx.agents.agent_search.deep_search_a.initial.generate_individual_sub_answer.states import (
     AnswerQuestionOutput,
 )
 from onyx.agents.agent_search.deep_search_a.main.operations import logger
