+++ backend/onyx/agents/agent_search/deep_search_a/main/edges.py
@@ -6,10 +6,10 @@ from typing import Literal
 from langchain_core.runnables import RunnableConfig
 from langgraph.types import Send
 
-from onyx.agents.agent_search.deep_search_a.initial.individual_sub_answer_generation.states import (
+from onyx.agents.agent_search.deep_search_a.initial.generate_individual_sub_answer.states import (
     AnswerQuestionInput,
 )
-from onyx.agents.agent_search.deep_search_a.initial.individual_sub_answer_generation.states import (
+from onyx.agents.agent_search.deep_search_a.initial.generate_individual_sub_answer.states import (
     AnswerQuestionOutput,
 )
 from onyx.agents.agent_search.deep_search_a.main.states import MainState
