+++ backend/onyx/agents/agent_search/deep_search_a/refinement/sub_answer_consolidation/edges.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 from langgraph.types import Send
 
-from onyx.agents.agent_search.deep_search_a.initial.individual_sub_answer_generation.states import (
+from onyx.agents.agent_search.deep_search_a.initial.generate_individual_sub_answer.states import (
     AnswerQuestionInput,
 )
 from onyx.agents.agent_search.deep_search_a.shared.expanded_retrieval.states import (
